move fade code into its own file
